クラス名をオブジェクト指向らしくするために
