--- 99.wc_object/wc_file.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-class WcFile
-  attr_reader :line_count, :word_count, :byte_count, :name
-
-  def initialize(text:, filename: '')
-    @line_count = text.split("\n").size
-    @word_count = text.split.size
-    @byte_count = text.bytesize
-    @name = filename
-  end
-end
